BYASP: experimental PDM plotting
This patch adds a function to plot the 2D and 3D landmarks.
The intent is to attempt and see how I can use this data
to extrapolate facial animation. I think I have enough data.
The challenge is going to be how transfer this data onto
an MB model (or another)

Signed-off-by: Amir Shehata <amir.shehata@gmail.com>